rename `meta_data` to `meta_snapshot` in manifest.
